Restored limits. Removed client-side database backup at startup Forced isARawProduct=0 everywhere. Hidden '*' special product.

+++ src/structs.h
@@ -126,7 +126,7 @@ struct ClientInfo
   QStringList   tags;
   QDate        lastCreditReset;
   // Effective limits for the current transaction
-  QHash<int, Limit>         limits;
+  QHash<qulonglong, Limit>         limits;
 };
